Implement data layer for food fetching

+++ app/src/main/java/uz/androdev/fooddelivery/model/Food.kt
@@ -8,9 +8,9 @@ package uz.androdev.fooddelivery.model
  */
 
 data class Food(
-    val id: Int,
+    val id: Long,
     val image: String,
     val title: String,
     val description: String,
-    val price: Float
+    val price: Int
 )
